# src/Pitchwire.Web/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Pitchwire.Api | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:5199/"
    }
  ],
  "paths": {
...
    "/devices/me/preferences": {
      "put": {
        "tags": [
          "DeviceEndpoints"
        ],
        "operationId": "UpdatePreferences",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreferencesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceSettings"
                }
              }
            }
          }
        }
      }
    },
    "/devices/me/favourites": {
      "put": {
        "tags": [
          "DeviceEndpoints"
        ],
        "operationId": "UpdateFavourites",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FavouritesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceSettings"
                }
              }
            }
          }
        }
      }
    },
    "/devices/me/push-subscriptions": {
      "post": {
        "tags": [
          "DeviceEndpoints"
        ],
        "operationId": "Subscribe",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubscriptionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      },
      "delete": {
        "tags": [
          "DeviceEndpoints"
        ],
        "operationId": "Unsubscribe",
        "parameters": [
          {
            "name": "endpoint",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
...
    "/seasons/{seasonId}/results": {
      "get": {
        "tags": [
          "ReadEndpoints"
        ],
        "operationId": "Results",
        "parameters": [
          {
            "name": "seasonId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "$top",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "$skiptoken",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageOfMatchSummary"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "DeviceSettings": {
        "required": [
          "id",
          "timeZoneId",
          "quietHoursStart",
          "quietHoursEnd",
          "notifyOnGoal",
          "notifyOnRedCard",
          "notifyOnKickoff",
          "notifyOnFullTime",
          "favourites"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "timeZoneId": {
            "type": "string"
          },
          "quietHoursStart": {
            "type": [
              "null",
              "string"
            ],
            "format": "time"
          },
          "quietHoursEnd": {
            "type": [
              "null",
              "string"
            ],
            "format": "time"
          },
          "notifyOnGoal": {
            "type": "boolean"
          },
          "notifyOnRedCard": {
            "type": "boolean"
          },
          "notifyOnKickoff": {
            "type": "boolean"
          },
          "notifyOnFullTime": {
            "type": "boolean"
          },
          "favourites": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          }
        }
      },
      "FavouritesRequest": {
        "required": [
          "teamIds"
        ],
        "type": "object",
        "properties": {
          "teamIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          }
        }
      },
...
      "MatchSummary": {
        "required": [
          "id",
          "round",
          "kickoffUtc",
          "status",
          "minute",
          "homeScore",
          "awayScore",
          "isDegraded",
          "home",
          "away"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "round": {
            "type": "integer",
            "format": "int32"
          },
          "kickoffUtc": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "string"
          },
          "minute": {
            "type": "integer",
            "format": "int32"
          },
          "homeScore": {
            "type": "integer",
            "format": "int32"
          },
          "awayScore": {
            "type": "integer",
            "format": "int32"
          },
          "isDegraded": {
            "type": "boolean"
          },
          "home": {
            "$ref": "#/components/schemas/TeamRef"
          },
          "away": {
            "$ref": "#/components/schemas/TeamRef"
          }
        }
      },
      "PageOfMatchSummary": {
        "required": [
          "value",
          "nextLink"
        ],
        "type": "object",
        "properties": {
          "value": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MatchSummary"
            }
          },
          "nextLink": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "PreferencesRequest": {
        "required": [
          "timeZoneId",
          "quietHoursStart",
          "quietHoursEnd",
          "notifyOnGoal",
          "notifyOnRedCard",
          "notifyOnKickoff",
          "notifyOnFullTime"
        ],
        "type": "object",
        "properties": {
          "timeZoneId": {
            "type": "string"
          },
          "quietHoursStart": {
            "type": [
              "null",
              "string"
            ],
            "format": "time"
          },
          "quietHoursEnd": {
            "type": [
              "null",
              "string"
            ],
            "format": "time"
          },
          "notifyOnGoal": {
            "type": "boolean"
          },
          "notifyOnRedCard": {
            "type": "boolean"
          },
          "notifyOnKickoff": {
            "type": "boolean"
          },
          "notifyOnFullTime": {
            "type": "boolean"
          }
        }
      },
...
      "SubscriptionKeys": {
        "required": [
          "p256dh",
          "auth"
        ],
        "type": "object",
        "properties": {
          "p256dh": {
            "type": "string"
          },
          "auth": {
            "type": "string"
          }
        }
      },
      "SubscriptionRequest": {
        "required": [
          "endpoint",
          "keys"
        ],
        "type": "object",
        "properties": {
          "endpoint": {
            "type": "string"
          },
          "keys": {
            "$ref": "#/components/schemas/SubscriptionKeys"
          }
        }
      },
...
      "TeamRef": {
        "required": [
          "id",
          "name",
          "shortName",
          "slug"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "shortName": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          }
        }
      }
    }
  },
...
}